Add Translate module
