+++ modules/__init__.py
@@ -1,2 +1,3 @@
 import modules.text2voice
+import modules.voice2voice
 import modules.insturment_resemble
