+++ xtts_webui.py
@@ -1,8 +1,4 @@
-from scripts.voice2voice import infer_rvc, get_openvoice_refs, infer_rvc_batch, infer_openvoice, find_openvoice_ref_by_name
-from scripts.funcs import save_audio_to_wav
-import shutil
-from datetime import datetime
-import glob
+
 from scripts.modeldownloader import get_folder_names_advanced
 from scripts.tts_funcs import TTSWrapper
 
@@ -52,7 +48,7 @@ supported_languages = {
 reversed_supported_languages = {
     name: code for code, name in supported_languages.items()}
 reversed_supported_languages_list = list(reversed_supported_languages.keys())
-
+supported_languages_list = list(supported_languages.keys())
 # INIT MODEL
 XTTS = TTSWrapper(OUTPUT_FOLDER, SPEAKER_FOLDER, LOWVRAM_MODE,
                   MODEL_SOURCE, MODEL_VERSION, DEVICE)
@@ -60,205 +56,7 @@ XTTS = TTSWrapper(OUTPUT_FOLDER, SPEAKER_FOLDER, LOWVRAM_MODE,
 # LOAD MODEL
 logger.info(f"Start loading model {MODEL_VERSION}")
 this_dir = Path(__file__).parent.resolve()
-# XTTS.load_model(this_dir)
-
-
-# Constants
-WAV_EXTENSION = "*.wav"
-MP3_EXTENSION = "*.mp3"
-FLAC_EXTENSION = "*.flac"
-
-DATE_FORMAT = "%Y%m%d_%H%M%S"
-SPEAKER_PREFIX = "speaker/"
-REFERENCE_KEYWORD = "reference"
-
-# Auxiliary functions
-
-
-def get_reference_path(speaker_wav, speaker_path_text):
-    if speaker_wav == REFERENCE_KEYWORD:
-        return speaker_path_text if speaker_path_text else None
-    else:
-        ref_path = XTTS.get_speaker_path(speaker_wav)
-        return ref_path[0] if isinstance(ref_path, list) else ref_path
-
-
-def find_audio_files(batch_path):
-    return glob.glob(os.path.join(batch_path, WAV_EXTENSION)) + \
-        glob.glob(os.path.join(batch_path, MP3_EXTENSION)) + \
-        glob.glob(os.path.join(batch_path, FLAC_EXTENSION))
-
-# Main optimization function
-
-
-def infer_openvoice_audio(openvoice_audio_single, openvoice_audio_batch, openvoice_audio_batch_path,
-                          openvoice_voice_ref_list, openvoice_status_bar, speaker_path_text):
-    print("hello world")
-
-    if openvoice_voice_ref_list == "None":
-        return None, None, "Please select Reference audio"
-
-    if not openvoice_audio_single and not openvoice_audio_batch and not openvoice_audio_batch_path:
-        return None, None, "Please load audio"
-
-    output_folder = this_dir / OUTPUT_FOLDER
-    folder_name = "openvoice_" + datetime.now().strftime(DATE_FORMAT)
-
-    # Save Audio
-    input_file = None
-    if openvoice_audio_single:
-        rate, y = openvoice_audio_single
-        input_file = save_audio_to_wav(rate, y, Path.cwd())
-
-    audio_files = openvoice_audio_batch or []
-    if openvoice_audio_batch_path:
-        audio_files.extend(find_audio_files(openvoice_audio_batch_path))
-
-    openvoice_status_bar = gr.Progress(track_tqdm=True)
-    if audio_files:
-        output_folder = output_folder / folder_name
-        os.makedirs(output_folder, exist_ok=True)
-        tqdm_object = openvoice_status_bar.tqdm(
-            audio_files, desc="Tuning Files...")
-
-        for audio_file in tqdm_object:
-            ref_voice_opvoice_path = None
-            allow_infer = True
-
-            if openvoice_voice_ref_list.startswith(SPEAKER_PREFIX):
-                speaker_wav = openvoice_voice_ref_list.split("/")[-1]
-                ref_voice_opvoice_path = get_reference_path(
-                    speaker_wav, speaker_path_text)
-
-                if not ref_voice_opvoice_path:
-                    allow_infer = False
-                    print("Reference not found")
-            else:
-                ref_voice_opvoice_path = find_openvoice_ref_by_name(
-                    Path.cwd(), openvoice_voice_ref_list)
-
-            if allow_infer and ref_voice_opvoice_path:
-                
-                output_filename = output_folder / \
-                    f"openvoice_{Path(audio_file).stem}.wav"
-                infer_openvoice(
-                    input_path=audio_file, ref_path=ref_voice_opvoice_path, output_path=output_filename)
-
-        return None, None, f"Files saved in {output_folder} folder"
-
-    elif openvoice_audio_single:
-        temp_dir = Path.cwd() / "output"
-        filename_openvoice = Path(ref_voice_opvoice_path).stem
-        output_filename = temp_dir / \
-            f"openvoice_{filename_openvoice}_{datetime.now().strftime(DATE_FORMAT)}.wav"
-
-        if openvoice_voice_ref_list.startswith(SPEAKER_PREFIX):
-            speaker_wav = openvoice_voice_ref_list.split("/")[-1]
-            ref_voice_opvoice_path = get_reference_path(
-                speaker_wav, speaker_path_text)
-        else:
-            ref_voice_opvoice_path = find_openvoice_ref_by_name(
-                Path.cwd(), openvoice_voice_ref_list)
-
-        if ref_voice_opvoice_path:
-            infer_openvoice(
-                input_path=input_file, ref_path=ref_voice_opvoice_path, output_path=output_filename)
-            output_audio = output_filename
-            done_message = "Done"
-        else:
-            output_audio = None
-            done_message = "Reference not found"
-
-        return None, gr.Audio(label="Result", value=output_audio), done_message
-
-    # If none of the conditions are met, return an error message
-    return None, None, "An unexpected error occurred during processing"
-
-
-# Main optimization function
-def infer_rvc_audio(
-        rvc_audio_single,
-        rvc_audio_batch,
-        rvc_audio_batch_path,
-        rvc_voice_settings_model_name,
-        rvc_voice_settings_model_path,
-        rvc_voice_settings_index_path,
-        rvc_voice_settings_pitch,
-        rvc_voice_settings_index_rate,
-        rvc_voice_settings_protect_voiceless,
-        rvc_voice_settings_method,
-        rvc_voice_filter_radius,
-        rvc_voice_resemple_rate,
-        rvc_voice_envelope_mix,
-        rvc_voice_status_bar
-):
-    if not rvc_voice_settings_model_name:
-        return None, None, "Please select RVC model"
-
-    if not (rvc_audio_single or rvc_audio_batch or rvc_audio_batch_path):
-        return None, None, "Please load audio"
-
-    output_folder = this_dir / OUTPUT_FOLDER
-    folder_name = f"rvc_{datetime.now().strftime(DATE_FORMAT)}"
-    done_message = ""
-
-    input_file = None
-    if rvc_audio_single:
-        rate, y = rvc_audio_single
-        input_file = save_audio_to_wav(rate, y, Path.cwd())
-
-    audio_files = rvc_audio_batch or []
-    if rvc_audio_batch_path:
-        audio_files.extend(find_audio_files(rvc_audio_batch_path))
-
-    rvc_voice_status_bar = gr.Progress(track_tqdm=True)
-    # Process batches of files
-    if audio_files:
-        output_folder = output_folder / folder_name / "temp"
-        os.makedirs(output_folder, exist_ok=True)
-
-        output_audio = infer_rvc_batch(
-            model_name=rvc_voice_settings_model_name,
-            pitch=rvc_voice_settings_pitch,
-            index_rate=rvc_voice_settings_index_rate,
-            protect_voiceless=rvc_voice_settings_protect_voiceless,
-            method=rvc_voice_settings_method,
-            index_path=rvc_voice_settings_model_path,
-            model_path=rvc_voice_settings_index_path,
-            paths=audio_files,
-            opt_path=output_folder,
-            filter_radius=rvc_voice_filter_radius,
-            resemple_rate=rvc_voice_resemple_rate,
-            envelope_mix=rvc_voice_envelope_mix,
-        )
-
-        done_message = f"Done, file saved in {folder_name} folder"
-        return None, None, done_message
-
-    # Process single file
-    elif rvc_audio_single:
-        output_file_name = output_folder / \
-            f"rvc_{rvc_voice_settings_model_name}_{datetime.now().strftime(DATE_FORMAT)}.wav"
-        infer_rvc(
-            pitch=rvc_voice_settings_pitch,
-            index_rate=rvc_voice_settings_index_rate,
-            protect_voiceless=rvc_voice_settings_protect_voiceless,
-            method=rvc_voice_settings_method,
-            index_path=rvc_voice_settings_model_path,
-            model_path=rvc_voice_settings_index_path,
-            input_path=input_file,
-            opt_path=output_file_name,
-            filter_radius=rvc_voice_filter_radius,
-            resemple_rate=rvc_voice_resemple_rate,
-            envelope_mix=rvc_voice_envelope_mix
-        )
-
-        done_message = "Done"
-        output_audio = output_file_name
-        return None, gr.Audio(label="Result", value=output_audio), done_message
-
-    # If none of the conditions are met, return an error message
-    return None, None, "An unexpected error occurred during processing"
+XTTS.load_model(this_dir)
 
 
 with gr.Blocks(css=css) as demo:
@@ -279,84 +77,7 @@ with gr.Blocks(css=css) as demo:
         from parts.text2voice import *
 
     with gr.Tab("Voice2Voice"):
-        with gr.Tab("RVC"):
-            with gr.Row():
-                with gr.Column():
-                    with gr.Tab("Single"):
-                        rvc_audio_single = gr.Audio(
-                            label="Single file", value=None)
-                    with gr.Tab("Batch"):
-                        rvc_audio_batch = gr.File(
-                            file_count="multiple", label="Batch files", file_types=["audio"])
-                        rvc_audio_batch_path = gr.Textbox(
-                            label="Path to folder with audio files (High priority)", value=None)
-                    with gr.Column():
-                        with gr.Row():
-                            rvc_voice_settings_model_name = gr.Dropdown(
-                                label="RVC Model name", info="Create a folder with your model name in the rvc folder and put .pth and .index there , .index optional", choices=rvc_models)
-                            rvc_voice_settings_update_btn = gr.Button(
-                                value="Update", elem_classes="rvc_update-btn", visible=True)
-                        rvc_voice_settings_model_path = gr.Textbox(
-                            label="RVC Model", value="", visible=True, interactive=False)
-                        rvc_voice_settings_index_path = gr.Textbox(
-                            label="Index file", value="", visible=True, interactive=False)
-                        rvc_voice_settings_pitch = gr.Slider(
-                            minimum=-24, maximum=24, value=0, step=1, label="Pitch")
-                        rvc_voice_settings_index_rate = gr.Slider(
-                            minimum=0, maximum=1, value=0.75, step=0.01, label="Index rate")
-                        rvc_voice_settings_protect_voiceless = gr.Slider(
-                            minimum=0, maximum=0.5, value=0.33, step=0.01, label="Protect voiceless")
-                        rvc_voice_settings_method = gr.Radio(
-                            ["crepe", "pm", "rmvpe", "harvest"], value="rmvpe", label="RVC Method")
-                        rvc_voice_filter_radius = gr.Slider(
-                            minimum=0, maximum=7, value=3, step=1, label="If >=3: apply median filtering to the harvested pitch results. The value represents the filter radius and can reduce breathiness.")
-                        rvc_voice_resemple_rate = gr.Slider(
-                            minimum=0, maximum=48000, value=0, step=1, label="Resample the output audio in post-processing to the final sample rate. Set to 0 for no resampling")
-                        rvc_voice_envelope_mix = gr.Slider(
-                            minimum=0, maximum=1, value=0.25, step=0.01, label="Use the volume envelope of the input to replace or mix with the volume envelope of the output. The closer the ratio is to 1, the more the output envelope is used")
-                with gr.Column():
-                    rvc_voice_status_bar = gr.Label(
-                        value="Choose model and click Infer button")
-                    rvc_video_output = gr.Video(
-                        label="Waveform RVC", value=None, interactive=False, visible=False)
-                    rvc_voice_output = gr.Audio(
-                        label="Result", value=None, interactive=False)
-                    rvc_voice_infer_btn = gr.Button(value="Infer")
-
-        with gr.Tab("OpenVoice"):
-            with gr.Row():
-                with gr.Column():
-                    with gr.Tab("Single"):
-                        openvoice_audio_single = gr.Audio(
-                            label="Single file", value=None)
-                    with gr.Tab("Batch"):
-                        openvoice_audio_batch = gr.File(
-                            file_count="multiple", label="Batch files", file_types=["audio"])
-                        openvoice_audio_batch_path = gr.Textbox(
-                            label="Path to folder with audio files (High priority)", value=None)
-                    with gr.Column():
-                        openvoice_voice_ref_list = get_openvoice_refs(this_dir)
-
-                        if len(openvoice_voice_ref_list) == 0:
-                            openvoice_voice_ref_list = ["None"]
-
-                        gr.Markdown(
-                            "**Add samples to the voice2voice/openvoice audio files folder or select from the reference speaker list**")
-                        with gr.Row():
-                            opvoice_voice_ref_list = gr.Dropdown(
-                                label="Reference sample", value=openvoice_voice_ref_list[0], choices=openvoice_voice_ref_list)
-                            opvoice_voice_show_speakers = gr.Checkbox(
-                                value=False, label="Show choises from the speakers folder")
-                with gr.Column():
-                    openvoice_status_bar = gr.Label(
-                        value="Choose reference and click Infer button")
-                    openvoice_video_output = gr.Video(
-                        label="Waveform RVC", value=None, interactive=False, visible=False)
-                    openvoice_voice_output = gr.Audio(
-                        label="Result", value=None, interactive=False)
-                    openvoice_voice_infer_btn = gr.Button(value="Infer")
-        with gr.Tab("Translate"):
-            gr.Markdown("WIP")
+        from parts.voice2voice import *
 
     with gr.Tab("Train"):
         from parts.train import *
@@ -364,42 +85,7 @@ with gr.Blocks(css=css) as demo:
     with gr.Tab("Instuments"):
         from parts.instuments import *
 
-    from modules.text2voice.voice2voice import select_rvc_model, update_rvc_model, update_openvoice_ref_list
-    rvc_voice_settings_model_name.change(fn=select_rvc_model, inputs=[rvc_voice_settings_model_name], outputs=[
-        rvc_voice_settings_model_path, rvc_voice_settings_index_path])
-    rvc_voice_settings_update_btn.click(fn=update_rvc_model, inputs=[rvc_voice_settings_model_name], outputs=[
-        rvc_voice_settings_model_name, rvc_voice_settings_model_path, rvc_voice_settings_index_path])
-
-    rvc_voice_infer_btn.click(fn=infer_rvc_audio, inputs=[
-        # INPUT
-        rvc_audio_single,
-        rvc_audio_batch,
-        rvc_audio_batch_path,
-        # PATH
-        rvc_voice_settings_model_name,
-        rvc_voice_settings_model_path,
-        rvc_voice_settings_index_path,
-        # SETTINGS
-        rvc_voice_settings_pitch,
-        rvc_voice_settings_index_rate,
-        rvc_voice_settings_protect_voiceless,
-        rvc_voice_settings_method,
-        rvc_voice_filter_radius,
-        rvc_voice_resemple_rate,
-        rvc_voice_envelope_mix,
-        # STATUS
-        rvc_voice_status_bar
-    ], outputs=[
-        rvc_video_output,
-        rvc_voice_output,
-        rvc_voice_status_bar
-    ])
-
-    opvoice_voice_show_speakers.change(fn=update_openvoice_ref_list, inputs=[
-        opvoice_voice_ref_list, opvoice_voice_show_speakers], outputs=[opvoice_voice_ref_list])
-
-    openvoice_voice_infer_btn.click(fn=infer_openvoice_audio, inputs=[openvoice_audio_single, openvoice_audio_batch, openvoice_audio_batch_path,
-                                    opvoice_voice_ref_list, openvoice_status_bar, speaker_path_text], outputs=[openvoice_video_output, openvoice_voice_output, openvoice_status_bar])
+    
     # LOAD FUNCTIONS AND HANDLERS
     import modules
 if __name__ == "__main__":
